- add thread tests - move basic_set to own file - fix make clean - reduce physical memory assigned to test vm to 2GB - add appropriate nobits / progbits section directives - always break into debugger on calls to validate_failed - do not continue on auto-load symbols (caused some problems with debugging) - add fpu zeroall compiled code callable helper - improve privilege check / RPL handling - validate flags on context restore and break into debugger - drastically increase heap validation - fix fsgsbase insn patching - fix tss code - add "any interrupt" hook function for debugging - many improvements in context save/restore/switch code paths - drastically reduce unnecessary fiddling with FPU - improve thread_info_t layout - fix fpu context save restore - debugging infrastructure to lookup interrupt handler names - add dbgout, putsdbg - export rand API - extensive improvements in set and vector unit tests

+++ libc/src/sys/module/init_module.cc
@@ -9,7 +9,7 @@ int init_module(void const *module,
                 ptrdiff_t module_sz, char const *module_name,
                 struct module *mod_user, const char *parameters)
 {
-    long err = syscall5(long(module), long(module_sz),
+    long err = syscall5(long(module), module_sz,
                         long(module_name), long(mod_user),
                         long(parameters),
                         SYS_init_module);
